Loading Models Faster, Screenshot added

+++ inc/stl.h
@@ -7,6 +7,7 @@ extern "C"
 #include <stdio.h>
 #include <stdlib.h>
 #include <stdbool.h>
+#include <string.h>
 #include <math.h>
 #include <errno.h>
 
@@ -18,7 +19,7 @@ extern int errno;
 #define STL_GL_10_BITS 10
 #define STL_GL_10_MAX ((2 << (STL_GL_10_BITS - 1))-1)
 
-int stl_model_init(const char *filename, float* vertices, size_t* vertexSize, int32_t* normals, size_t *normalSize);
+int stl_model_init(const char *filename, float* vertices, size_t* vertexSize, float* normals, size_t *normalSize);
 
 #ifdef __cplusplus
 } // extern "C"
